[DEV] Refactor/home (#20)
* chore(components): Delete moved file components

* refactor(components): Move and group each components into ui's folder

* chore(config): Modify next config for firebase storage

* feat(sidebar): Add sidebar component

* feat(module): Add data for destination location and destination type

* refactor(components): Update invoked components

* feat(hooks): Create custom hooks for destinations

* refactor(home): Refactored sidebar, hooks, and card component

* chore(partial): Adjust style and add missing props on recommendation section
