+++ app/(route)/landing/_partials/recomendation/page.tsx
@@ -1,5 +1,5 @@
 import { NextPage } from 'next'
-import CustomCard from '@/components/ui/CustomCard'
+import CustomCard from '@/components/ui/card/CustomCard'
 import { imageCards } from '@/data/index'
 interface Props {}
 
@@ -10,7 +10,7 @@ const RecommendationSection: NextPage<Props> = () => {
         <h1 className='py-12 text-center text-4xl font-medium'>
           What We Reccomends <br /> For You
         </h1>
-        <div className='grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4'>
+        <div className='grid grid-cols-1 justify-items-center gap-6 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4'>
           {imageCards.map(item => (
             <CustomCard
               key={item.id}
@@ -19,6 +19,7 @@ const RecommendationSection: NextPage<Props> = () => {
               rating={item.rating}
               location={item.location}
               description={item.description}
+              isSaveAvailable={false}
             />
           ))}
         </div>
